App Part/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-telegram-bot
httpx
